requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
openai==0.27.2
pip==22.3.1
python-dotenv==1.0.0
requests==2.28.2
